Caching Data In Room Done

+++ app/src/main/java/com/abdallah/prayerapp/data/model/PrayerNetworkModel.kt
@@ -115,6 +115,6 @@ data class Location(
 )
 
 data class Params(
-    val Fajr: Int,
-    val Isha: Int
+    val Fajr: Double,
+    val Isha: Double
 )
